Crate pallet_bridge_registry

Bridge Registry Module

A module for maintaining bridge metadata or views over connected sets of anchors.

Overview

The Bridge Registry module provides functionality maintaing and storing metadata about existing bridges.

The supported dispatchable functions are documented in the Call enum.

Terminology

Goals

Interface

Related Modules

System

Support

Re-exports

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

types