Trait pallet_bridge_registry::pallet::Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>;
    type BridgeIndex: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen;
    type MaxAdditionalFields: Get<u32> + MaybeSerializeDeserialize;
    type MaxResources: Get<u32> + MaybeSerializeDeserialize;
    type MaxProposalLength: Get<u32>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait. The module configuration trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The overarching RuntimeEvent type.

type ForceOrigin: EnsureOrigin<Self::RuntimeOrigin>

The origin which may forcibly reset parameters or otherwise alter privileged attributes.

type BridgeIndex: Encode + Decode + Parameter + AtLeast32Bit + Default + Copy + MaxEncodedLen

Bridge index type

type MaxAdditionalFields: Get<u32> + MaybeSerializeDeserialize

Maximum number of additional fields that may be stored in a bridge’s metadata. Needed to bound the I/O required to access an identity, but can be pretty high.

type MaxResources: Get<u32> + MaybeSerializeDeserialize

Maximum number of resources that may be stored in a bridge. This is not to be confused with the actual maximum supported by the bridge. Needed to bound the I/O required to access a metadata object, but can be pretty high.

type MaxProposalLength: Get<u32>

Max length of a proposal

type WeightInfo: WeightInfo

Weight information for the extrinsics